requirements.txt
openai
httpx
python-dotenv
pyyaml
fastapi
uvicorn
websockets
